+++ confidant/routes/credentials.py
@@ -25,7 +25,8 @@ from confidant.services import (
     webhook,
 )
 from confidant.services.ciphermanager import CipherManager
-from confidant.utils import maintenance, misc, stats
+from confidant.utils import maintenance, misc
+from confidant.utils.stats import stats
 from confidant.utils.dynamodb import decode_last_evaluated_key
 
 logging = importlib.import_module(settings.LOGGING_MODULE)
