--- confidant/utils/__init__.py
@@ -1,9 +1,0 @@
-import statsd
-
-from confidant import settings
-
-stats = statsd.StatsClient(
-    settings.STATSD_HOST,
-    settings.STATSD_PORT,
-    prefix='confidant'
-)
