+++ confidant/utils/stats.py
@@ -1,0 +1,9 @@
+import statsd
+
+from confidant import settings
+
+stats = statsd.StatsClient(
+    settings.STATSD_HOST,
+    settings.STATSD_PORT,
+    prefix='confidant'
+)
